refactor(anvil): isolate Monad backend helpers
